Fix integer division Python 2 compatibility, TypeError: 'float' object is not iterable
28 / 8 in Python 2 is 3 (integer), but in Python 3 is 3.5 (float),
and type(math.ceil(3.5)) is float in Python 2 but int in Python 3.
Force floating-point division and add integer cast so it works in both.
Fixes this Python 2.7.12 error when disassembling:

Traceback (most recent call last):
  File "dis.py", line 281, in <module>
    ib = InputBuffer(f, file_len, bounds)
  File "dis.py", line 116, in __init__
    self.access = bytearray(math.ceil(available / 8))
TypeError: 'float' object is not iterable

+++ dis.py
@@ -113,7 +113,7 @@ class InputBuffer:
                 
         self.buffer = bytearray(available)
         # Stores which bytes have already been read
-        self.access = bytearray(math.ceil(available / 8))
+        self.access = bytearray(int(math.ceil(available / 8.0)))
         data.readinto(self.buffer)
         
     def was_read(self, o):
